renamed components, added notification store, fixed radio input, separated form component
