+++ .env
@@ -1,1 +1,2 @@
 BASE_URL = '/GTI-connect-frontend/'
+VUE_APP_ACCOUNT_API_URL = http://127.0.0.1:8000/api/accounts/
