+++ src/components/molecules/NotificationContainer.vue
@@ -1,0 +1,28 @@
+<template>
+  <div class="notification-container">
+    <notification v-for="notification in notifications" :key="notification.id" :notification="notification" />
+  </div>
+</template>
+
+<script>
+import Notification from '../atoms/Notification.vue'
+import { mapState } from 'vuex'
+  export default {
+    components: { Notification },
+    computed: mapState('notification', ['notifications'])
+  }
+</script>
+
+<style lang="scss" scoped>
+.notification-container {
+  position: fixed;
+  top: 90px;
+  right: 10px;
+  padding-right: 10px;
+  padding-left: 10px;
+  background-color: var(--bgcolour);
+  border-radius: 12px;
+  box-shadow: 0px 0px 15px -8px var(--boxshadowcolour);
+}
+
+</style>
