+++ src/services/UserService.js
@@ -1,7 +1,7 @@
 import axios from 'axios'
 
 const apiClient = axios.create({
-  baseURL: 'http://127.0.0.1:8000/api/accounts/',
+  baseURL: process.env.VUE_APP_ACCOUNT_API_URL,
   timeout: 10000
 })
