+++ src/store/index.js
@@ -1,6 +1,6 @@
 import { createStore } from 'vuex'
 import * as user from '@/store/modules/user.js'
-
+import * as notification from '@/store/modules/notification.js'
 export default createStore({
   state: {
   },
@@ -9,6 +9,7 @@ export default createStore({
   actions: {
   },
   modules: {
-    user
+    user,
+    notification
   }
 })
