+++ src/store/modules/notification.js
@@ -1,0 +1,30 @@
+export const namespaced = true
+
+export const state = {
+  notifications: []
+}
+
+let notiId = 1
+
+export const mutations = {
+  PUSH(state, notification) {
+    state.notifications.push({
+      ...notification,
+      id: notiId++
+    })
+  },
+  DELETE(state, notificationToRemove) {
+    state.notifications = state.notifications.filter(
+      notification => notification.id !== notificationToRemove.id
+    )
+  }
+}
+
+export const actions = {
+  add({ commit }, notification) {
+    commit('PUSH', notification)
+  },
+  remove({ commit }, notificationToRemove) {
+    commit('DELETE', notificationToRemove)
+  }
+}
