+++ vue.config.js
@@ -15,5 +15,5 @@ module.exports = {
     devServer: {
       host: 'localhost',
       port: '3000'
-    }
+    },
 }
